PostProccess Now Longer Receives Modules
We will be removing esperanto and slightly modify the flow of trees.
PostProcess hook will now recieve a fully preprocessed tree instead of
one that was only partially processed.

While these changes have been verfiied in a real app, we loose the
ability to throw when a dev attempts to use a named import.